Update AcesUp.flt.html
I basically tabbed everything out and made it more readable. Kevin, could you take a look at the java part and see if it will run?

Also changed CSS so it is in order and easier to read.

I left the old java code in there and my suggested code as a comment. To upload new java, simply delete old code and un-comment the new code then, save.